+++ installing/deps-restorer/src/index.ts
@@ -413,6 +413,7 @@ export async function headlessInstall (opts: HeadlessOptions): Promise<Installat
         lockfileDir: opts.lockfileDir,
         preferSymlinkedExecutables: opts.preferSymlinkedExecutables,
         sideEffectsCacheRead: opts.sideEffectsCacheRead,
+        supportedArchitectures: opts.supportedArchitectures,
       })
       stageLogger.debug({
         prefix: lockfileDir,
@@ -451,6 +452,7 @@ export async function headlessInstall (opts: HeadlessOptions): Promise<Installat
           lockfileDir: opts.lockfileDir,
           sideEffectsCacheRead: opts.sideEffectsCacheRead,
           storeDir: opts.storeDir,
+          supportedArchitectures: opts.supportedArchitectures,
         }),
       ])
     }
@@ -889,6 +891,7 @@ async function linkAllPkgs (
     lockfileDir: string
     sideEffectsCacheRead: boolean
     storeDir: string
+    supportedArchitectures?: SupportedArchitectures
   }
 ): Promise<void> {
   // Create a marker source file that will be added to filesMap for GVS packages
@@ -917,6 +920,7 @@ async function linkAllPkgs (
           sideEffectsCacheKey = calcDepState(opts.depGraph, opts.depsStateCache, depNode.dir, {
             includeDepGraphHash: !opts.ignoreScripts && depNode.requiresBuild, // true when is built
             patchFileHash: depNode.patch?.hash,
+            supportedArchitectures: opts.supportedArchitectures,
           })
         }
       }