+++ __utils__/assert-project/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @pnpm/assert-project
 
+## 1100.0.1
+
+### Patch Changes
+
+- Updated dependencies [ff28085]
+  - @pnpm/types@1101.0.0
+  - @pnpm/installing.modules-yaml@1100.0.1
+  - @pnpm/lockfile.types@1100.0.1
+  - @pnpm/assert-store@1100.0.1
+
 ## 1001.0.0
 
 ### Major Changes