+++ fetching/fetcher-base/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @pnpm/fetcher-base
 
+## 1100.0.1
+
+### Patch Changes
+
+- Updated dependencies [ff28085]
+  - @pnpm/types@1101.0.0
+  - @pnpm/resolving.resolver-base@1100.0.1
+
 ## 1002.0.0
 
 ### Major Changes