+++ hooks/pnpmfile/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @pnpm/pnpmfile
 
+## 1100.0.1
+
+### Patch Changes
+
+- Updated dependencies [ff28085]
+  - @pnpm/types@1101.0.0
+  - @pnpm/core-loggers@1100.0.1
+  - @pnpm/hooks.types@1100.0.1
+  - @pnpm/lockfile.types@1100.0.1
+  - @pnpm/store.controller-types@1100.0.1
+  - @pnpm/crypto.hash@1100.0.0
+
 ## 1003.0.0
 
 ### Major Changes