+++ installing/linking/modules-cleaner/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @pnpm/modules-cleaner
 
+## 1100.0.1
+
+### Patch Changes
+
+- Updated dependencies [ff28085]
+  - @pnpm/types@1101.0.0
+  - @pnpm/bins.remover@1100.0.1
+  - @pnpm/core-loggers@1100.0.1
+  - @pnpm/deps.path@1100.0.1
+  - @pnpm/lockfile.filtering@1100.0.1
+  - @pnpm/lockfile.types@1100.0.1
+  - @pnpm/lockfile.utils@1100.0.1
+  - @pnpm/store.controller-types@1100.0.1
+
 ## 1002.0.0
 
 ### Major Changes