+++ lockfile/utils/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @pnpm/lockfile-utils
 
+## 1100.0.1
+
+### Patch Changes
+
+- Updated dependencies [ff28085]
+  - @pnpm/types@1101.0.0
+  - @pnpm/deps.path@1100.0.1
+  - @pnpm/hooks.types@1100.0.1
+  - @pnpm/lockfile.types@1100.0.1
+  - @pnpm/resolving.resolver-base@1100.0.1
+  - @pnpm/fetching.pick-fetcher@1100.0.1
+
 ## 1004.0.0
 
 ### Major Changes