+++ store/create-cafs-store/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @pnpm/create-cafs-store
 
+## 1100.0.1
+
+### Patch Changes
+
+- @pnpm/building.pkg-requires-build@1100.0.1
+- @pnpm/fetching.fetcher-base@1100.0.1
+- @pnpm/store.cafs@1100.0.1
+- @pnpm/store.controller-types@1100.0.1
+- @pnpm/fs.indexed-pkg-importer@1100.0.1
+
 ## 1001.0.0
 
 ### Major Changes