+++ workspace/project-manifest-reader/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @pnpm/read-project-manifest
 
+## 1100.0.2
+
+### Patch Changes
+
+- Updated dependencies [ff7733c]
+  - @pnpm/pkg-manifest.utils@1100.1.0
+
+## 1100.0.1
+
+### Patch Changes
+
+- Updated dependencies [ff28085]
+  - @pnpm/types@1101.0.0
+  - @pnpm/pkg-manifest.utils@1100.0.1
+  - @pnpm/workspace.project-manifest-writer@1100.0.1
+
 ## 1002.0.0
 
 ### Major Changes